+++ drivers/net/ppp/ppp_synctty.c
@@ -324,7 +324,7 @@ ppp_synctty_ioctl(struct tty_struct *tty, unsigned int cmd, unsigned long arg)
 /* May sleep, don't call from interrupt level or with interrupts disabled */
 static void
 ppp_sync_receive(struct tty_struct *tty, const unsigned char *buf,
-		  const char *cflags, int count)
+		 const char *cflags, size_t count)
 {
 	struct syncppp *ap = sp_get(tty);
 	unsigned long flags;
